fix #55: recommends the use of the official python3-pefile

+++ requirements.sh
@@ -16,6 +16,12 @@ if [ $PYTHON_VERSION -ne 0 ]; then
     exit
 fi
 
+python3 -c 'import future' >/dev/null 2>/dev/null
+if [ $? -ne 0 ]; then
+    echo "error: unable to find the package 'future' for python3 (required for pefile)"
+    exit
+fi
+
 rm -rf /usr/lib/python3.*/site-packages/capstone*
 rm -rf build
 
@@ -37,9 +43,9 @@ cd ../../..
 
 # PE
 # TODO: pip3 install pefile
-# -> fails due to a syntax error
-git clone -b master --depth 1 https://github.com/mlaferrera/python3-pefile
-cd python3-pefile
+# -> it fails due to a syntax error
+git clone -b master --depth 1 https://github.com/erocarrera/pefile
+cd pefile
 sudo -H python3 setup.py install
 cd ..
